Refactored group-address parsing functionality to be in shelley blockchain utils, instead of DB
