+++ src/entities/postgres-storage/db-shelley.js
@@ -2,6 +2,7 @@
 
 import _ from 'lodash'
 
+import { shelleyUtils } from '../../blockchain/shelley'
 import { CERT_TYPE } from '../../blockchain/shelley/certificate'
 import type {
   ShelleyTxType as TxType,
@@ -183,47 +184,10 @@ class DBShelley extends DB<TxType> implements Database<TxType> {
   }
 
   getGroupAddressesData(txDbData: TxDbDataType) {
-    const wasm = global.jschainlibs
-    const {
-      inputAddresses, outputAddresses,
-    } = txDbData
+    const { inputAddresses, outputAddresses } = txDbData
     const allAddresses = _.uniq([...inputAddresses, ...outputAddresses])
     this.logger.debug(`metadataCreator.allAddresses = ${allAddresses}`)
-    return allAddresses.map(addressString => {
-      let address
-      try {
-        address = wasm.Address.from_bytes(Buffer.from(addressString, 'hex'))
-      } catch (e) {
-        const prefix = addressString.substring(0, 3)
-        // TODO: find a better way to distinguish legacy funds?
-        if (prefix !== 'Ddz' && prefix !== 'Ae2') {
-          throw new Error(`Group Metadata could not parse address: ${addressString}`)
-        }
-        return null
-      }
-      const groupAddress = address.to_group_address()
-      let result = null
-      if (groupAddress) {
-        const spendingKey = groupAddress.get_spending_key()
-        const accountKey = groupAddress.get_account_key()
-        const discrim = address.get_discrimination()
-        const singleAddress = wasm.Address.single_from_public_key(spendingKey, discrim)
-        const accountAddress = wasm.Address.account_from_public_key(accountKey, discrim)
-        const metadata = {
-          groupAddress: addressString,
-          utxoAddress: Buffer.from(singleAddress.as_bytes()).toString('hex'),
-          accountAddress: Buffer.from(accountAddress.as_bytes()).toString('hex'),
-        }
-        singleAddress.free()
-        accountAddress.free()
-        spendingKey.free()
-        accountKey.free()
-        groupAddress.free()
-        result = metadata
-      }
-      address.free()
-      return result
-    }).filter(Boolean)
+    return allAddresses.map(shelleyUtils.splitGroupAddress).filter(x => x.groupAddress)
   }
 
   async getTxInputsDbData(tx: TxType, txUtxos: Array<mixed> = []): Promise<TxInputsDbDataType> {
